Move CommandLine from starboard/shared to starboard/common
b/413436961
