+++ starboard/android/shared/BUILD.gn
@@ -91,8 +91,6 @@ static_library("starboard_platform") {
     "//starboard/shared/starboard/audio_sink/audio_sink_is_valid.cc",
     "//starboard/shared/starboard/audio_sink/stub_audio_sink_type.cc",
     "//starboard/shared/starboard/audio_sink/stub_audio_sink_type.h",
-    "//starboard/shared/starboard/command_line.cc",
-    "//starboard/shared/starboard/command_line.h",
     "//starboard/shared/starboard/decode_target/decode_target_get_info.cc",
     "//starboard/shared/starboard/decode_target/decode_target_internal.h",
     "//starboard/shared/starboard/decode_target/decode_target_release.cc",
