+++ starboard/common/BUILD.gn
@@ -33,6 +33,8 @@ static_library("common") {
     "//starboard/shared/media_session/playback_state.h",
     "allocator.cc",
     "allocator.h",
+    "command_line.cc",
+    "command_line.h",
     "common.cc",
     "condition_variable.cc",
     "condition_variable.h",
