+++ starboard/linux/x64x11/main.cc
@@ -21,9 +21,9 @@
 #include "starboard/shared/signal/debug_signals.h"
 #include "starboard/shared/signal/suspend_signals.h"
 #if SB_IS(EVERGREEN_COMPATIBLE)
+#include "starboard/common/command_line.h"
 #include "starboard/common/paths.h"
 #include "starboard/elf_loader/elf_loader_constants.h"
-#include "starboard/shared/starboard/command_line.h"
 #endif
 #include "starboard/shared/x11/application_x11.h"
 
@@ -39,7 +39,7 @@ extern "C" SB_EXPORT_PLATFORM int main(int argc, char** argv) {
   starboard::shared::signal::InstallSuspendSignalHandlers();
 
 #if SB_IS(EVERGREEN_COMPATIBLE)
-  auto command_line = starboard::shared::starboard::CommandLine(argc, argv);
+  auto command_line = starboard::CommandLine(argc, argv);
   auto evergreen_content_path =
       command_line.GetSwitchValue(starboard::elf_loader::kEvergreenContent);
   std::string ca_certificates_path =
