+++ starboard/shared/starboard/application.h
@@ -23,6 +23,7 @@
 #include <memory>
 #include <vector>
 
+#include "starboard/common/command_line.h"
 #include "starboard/common/condition_variable.h"
 #include "starboard/common/log.h"
 #include "starboard/common/ref_counted.h"
@@ -30,7 +31,6 @@
 #include "starboard/event.h"
 #include "starboard/player.h"
 #include "starboard/shared/internal_only.h"
-#include "starboard/shared/starboard/command_line.h"
 #include "starboard/shared/starboard/player/filter/video_frame_internal.h"
 #include "starboard/types.h"
 #include "starboard/window.h"
